Update on "Fuzzing benchmark for FFT operators"
[ghstack-poisoned]
